+++ src/main/java/nl/openminetopia/modules/player/utils/PlaytimeUtil.java
@@ -6,11 +6,11 @@ import nl.openminetopia.configuration.MessageConfiguration;
 @UtilityClass
 public class PlaytimeUtil {
 
-    public static String formatPlaytime(int playtimeInSeconds) {
-        int days = playtimeInSeconds / 86400;
-        int hours = (playtimeInSeconds % 86400) / 3600;
-        int minutes = ((playtimeInSeconds % 86400) % 3600) / 60;
-        int seconds = ((playtimeInSeconds % 86400) % 3600) % 60;
+    public static String formatPlaytime(long playtimeInSeconds) {
+        long days = playtimeInSeconds / 86400;
+        long hours = (playtimeInSeconds % 86400) / 3600;
+        long minutes = ((playtimeInSeconds % 86400) % 3600) / 60;
+        long seconds = ((playtimeInSeconds % 86400) % 3600) % 60;
         return MessageConfiguration.message("player_time_format")
                 .replace("<days>", days + "")
                 .replace("<hours>", hours + "")
